refactor: schedule 도메인 분리 - 생각보다 너무 커져서 Farm 안에 집어 넣어버렸습니다 #13
