+++ src/main/java/poomasi/domain/farm/_schedule/dto/FarmScheduleRequest.java
@@ -1,4 +1,4 @@
-package poomasi.domain.farm.dto;
+package poomasi.domain.farm._schedule.dto;
 
 public record FarmScheduleRequest(
         Long farmId,
