+++ src/main/java/poomasi/domain/farm/_schedule/dto/FarmScheduleResponse.java
@@ -1,8 +1,8 @@
-package poomasi.domain.farm.dto;
+package poomasi.domain.farm._schedule.dto;
 
 import lombok.Builder;
-import poomasi.domain.farm.entity.FarmSchedule;
-import poomasi.domain.farm.entity.ScheduleStatus;
+import poomasi.domain.farm._schedule.entity.FarmSchedule;
+import poomasi.domain.farm._schedule.entity.ScheduleStatus;
 
 import java.time.LocalDate;
